+++ sopa/patches/__init__.py
@@ -1,1 +1,1 @@
-from .patches import Patches2D, BaysorPatches, _get_cell_id
+from .patches import Patches2D, _get_cell_id